Intial wiring of deposit into http endopoints, ITs

+++ package-ingest-impl/src/main/java/org/dataconservancy/packaging/impl/DcsPackageAnalyzer.java
@@ -37,7 +37,6 @@ import java.util.Map;
 import org.dataconservancy.packaging.ingest.PackageAnalyzer;
 import org.dataconservancy.packaging.ingest.PackagedResource;
 
-import org.apache.jena.atlas.RuntimeIOException;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.Property;
@@ -45,12 +44,16 @@ import org.apache.jena.rdf.model.RDFNode;
 import org.apache.jena.rdf.model.ResIterator;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.rdf.model.ResourceFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * @author bbrosius@jhu.edu
  */
 public class DcsPackageAnalyzer
         implements PackageAnalyzer {
+
+    static final Logger LOG = LoggerFactory.getLogger(DcsPackageAnalyzer.class);
 
     public static final String PARAM_EXTRACT_DIR = "pkg.extract.dir";
 
@@ -147,7 +150,7 @@ public class DcsPackageAnalyzer
             }
             // Read through the REM File to get the ldp concepts to populate the ldpresources
         } catch (final IOException e) {
-            throw new RuntimeIOException("Failed to open that package to retrieve the bag-info file. " + e
+            throw new RuntimeException("Failed to open that package to retrieve the bag-info file. " + e
                     .getMessage(), e);
         }
 
@@ -309,11 +312,14 @@ public class DcsPackageAnalyzer
 
     @Override
     public void cleanUpExtractionDirectory() {
-        try {
-            org.apache.commons.io.FileUtils.deleteDirectory(extractedPackageLocation);
-        } catch (final IOException e) {
-            throw new RuntimeIOException("Unable to clean up extract directory.", e);
+        if (extractedPackageLocation != null && extractedPackageLocation.exists()) {
+            try {
+                org.apache.commons.io.FileUtils.deleteDirectory(extractedPackageLocation);
+            } catch (final IOException e) {
+                throw new RuntimeException("Unable to clean up extract directory.", e);
+            }
+        } else {
+            LOG.info("No extraction directory to clean up");
         }
-
     }
 }
